requirements.txt
-e git+https://github.com/jim-bo/oncotreenx.git#egg=Oncotree
-e git+https://github.com/dfci/matchminer-engine.git@master#egg=matchengine
appnope==0.1.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
Cerberus==0.9.2
certifi==2016.2.28
colorlover==0.2.1
cufflinks==0.7.1
decorator==4.0.6
defusedxml==0.4.1
Eve==0.6.1
Eve-Swagger==0.0.9
Events==0.2.1
Fabric==1.10.2
Flask==0.10.1
Flask-Cors==2.1.2
Flask-PyMongo==0.3.1
gnureadline==6.3.3
google-api-python-client==1.5.0
gspread==0.3.0
isodate==0.5.0
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
j2cli>=0.3
lxml==4.1.1
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.1
networkx==1.10
nose==1.3.7
oauth2client==2.0.2
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
pkgconfig==1.1.0
plotly==1.9.3
ptyprocess==0.5
Pygments==2.0.2
PyYAML==3.11
pymongo==2.9.1
pyOpenSSL==17.5.0
python-dateutil==2.7.3
python-saml==2.1.5
pytz==2015.7
pyzmq==15.1.0
requests==2.9.1
setuptools==38.4.0
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
suds-jurko==0.6
terminado==0.5
tornado==4.3
traitlets==4.1.0b1
urllib3==1.15.1
Werkzeug==0.10.4
wheel==0.24.0
